% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otr.R
\name{otr}
\alias{otr}
\title{2D reconstruction}
\usage{
otr(points, masses = NULL, steps = NA, np = NA, tolerance = NA)
}
\arguments{
\item{points}{matrix of points, with two columns}

\item{masses}{numeric vector of point masses; if \code{NULL}, all masses are
set to \code{1}}

\item{steps}{number of edge collapse operators to be performed; if given
(i.e. not \code{NA}), then \code{np} and \code{tolerance} must be \code{NA}}

\item{np}{number of vertices to reach in the output; if given (i.e. not
\code{NA}), then \code{steps} and \code{tolerance} must be \code{NA}}

\item{tolerance}{tolerance on the Wasserstein distance; if given
(i.e. not \code{NA}), then \code{steps} and \code{np} must be \code{NA}}
}
\value{
A list containing two matrices: \code{vertices}, the isolated
  vertices, and \code{segments}.
}
\description{
Reconstruction of a 2D shape from a set of points.
}
\details{
See \href{https://doc.cgal.org/latest/Optimal_transportation_reconstruction_2/index.html}{Optimal Transportation Curve Reconstruction}.
}
\note{
Only one of \code{steps}, \code{np} and \code{tolerance} must be given.
}
\examples{
# reconstruct the pentagram
pts <- noisyPentagram()
shape <- otr(pts, np = 10)
segs <- shape[["segments"]]
# plot
opar <- par(mar = c(0, 0, 0, 0))
plot(
  pts, type = "p", pch = ".", asp = 1, cex = 4,
  xlab = NA, ylab = NA, axes = FALSE, col = "blue"
)
segments(
  segs[, "x0"], segs[, "y0"], segs[, "x1"], segs[, "y1"],
  lwd = 3, col = "black"
)
par(opar)
}
